Compatibility with Python 3
This only makes the necessary changes for the script to work on
Python 3. Future commits would make the script more readable and
"stable".

The script by default still won't run on Python 3. An environment
variable `SEEDMINER_ALLOW3` must be set to 1 to enable this for testing.
It should be removed once the script is confirmed to be stable on 2.7
and 3.x.

Output is the same on 2.7 and 3.x. Tested on 3.6.4.