Interactive menu added
When the script is run without any arguments, it will show a menu where you can choose what to do. The menu allows for CPU bruteforcing, GPU bruteforcing (with or without resuming from an offset) and Mii bruteforcing.
After bruteforcing, it will then ask if you want to delete no longer needed files (_part1.sed & backup, _part2.sed and input.bin) and if you want to rename the movable.sed to contain the FC
Also the script will now ignore if the ID0 is already in the movable_part1.sed file, since it works anyway.